migrate test camparison helpers
